--- differint/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ['GL','GLcoeffs','functionCheck','checkValues','isInteger','GLI','GLpoint','RLcoeffs','RLmatrix','RL','poch','RLpoint']
