+++ differintP/__init__.py
@@ -1,0 +1,15 @@
+from .core import *
+__all__ = [
+    "GL",
+    "GLcoeffs",
+    "functionCheck",
+    "checkValues",
+    "isInteger",
+    "GLI",
+    "GLpoint",
+    "RLcoeffs",
+    "RLmatrix",
+    "RL",
+    "poch",
+    "RLpoint",
+]
