+++ pyproject.toml
@@ -1,0 +1,43 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "differintP"
+version = "0.1.0"
+description = "Modern, pure Python fractional calculus library (fork of differint)"
+readme = "README.md"
+requires-python = ">=3.7"
+license = { file = "LICENSE" }
+authors = [
+    { name = "Your Name", email = "your.email@example.com" }
+]
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+    "Intended Audience :: Science/Research",
+    "Topic :: Scientific/Engineering :: Mathematics",
+]
+dependencies = [
+    # Add your runtime dependencies here, example:
+    "numpy >= 1.19",
+    "scipy >= 1.15.3"
+]
+
+[project.optional-dependencies]
+gpu = [
+    "cupy"
+]
+
+[project.urls]
+Homepage = "https://github.com/iparsw/differintP"
+Repository = "https://github.com/iparsw/differintP"
+Issues = "https://github.com/iparsw/differintP/issues"
+
+[tool.setuptools]
+package-dir = {"" = "differintP"}
+
+[tool.setuptools.packages.find]
+where = ["differintP"]
+
